@NickLaMuro 2018-06-22
@jrafanie yeah, and I am aware, and unsure what the issue is... still looking into it...

* * *

I had to make a change yesterday because `va_sql_cols` didn't exist:

https://github.com/ManageIQ/manageiq/pull/17626/files#diff-998021f35f1a1483d527b3ac70668c40


And the other conflict was because https://github.com/ManageIQ/manageiq/pull/17189 isn't in Gaprindashvili, so I just had to re-org the second commit.