@NickLaMuro 2018-06-22
@simaishi Looks like we need to backport a few things on both `gapindashvili` for the tests to pass:

On `gapindashvili`, just the following (in order, and there should be no conflicts):

* https://github.com/ManageIQ/manageiq/pull/16884
* https://github.com/ManageIQ/manageiq/pull/16931

On `fine`, we need the following (in order, and there will be conflicts):

* https://github.com/ManageIQ/manageiq/pull/14556
* https://github.com/ManageIQ/manageiq/pull/16884
* https://github.com/ManageIQ/manageiq/pull/16931

There will still be a conflict because https://github.com/ManageIQ/manageiq/pull/16915 was backported with conflicts and resolved by you manually (see [`gaprindashvili` backport](https://github.com/ManageIQ/manageiq/commit/01a01787c8c75d76ebf2d6f84d0b6ffdb9678021#diff-e5855dfa8cc9aeeeb9d13acd1525499eR479) v.s. [`fine` backport](https://github.com/ManageIQ/manageiq/commit/a6d06865680a9741aa22ee35d21227d86c01077e#diff-e5855dfa8cc9aeeeb9d13acd1525499eR809).

Besides that, should be pretty straight forward (I hope).

* * *

@kbrock you were involved, if not the author, of all of these PRs originally.  Can you do a risk assessment on all of them and tell us what is affected by these changes?  From what I can tell, basically all `MiqExpression` based filtering is affected, which is a lot.