@NickLaMuro 2018-06-22
Just talked with @simaishi on this one, and we are going to hold of on backporting this PR while I try an idea to rework the code to get the same results in a non-virtual attribute fashion specifically targeted at report columns.

Will leave this open for now while I work that up.