@kbrock 2018-06-22
@NickLaMuro `attribute_supported_by_sql?` has been improved for virtual attributes - so that will be necessary