@NickLaMuro 2018-06-22
@kbrock yeah, the work around I am thinking of forcing the backport is more surgical and will do the same thing as the virtual_attributes, but will take advantage of :extra_cols in RBac directly instead.

Won’t be as flexible, but that is the point, and will be less risky.