@NickLaMuro 2018-06-22
Okay, looks like my idea should work.  Going to open up a new PR for that one.