@kbrock 2018-06-29
NOTE: see #17629 for the g branch rework